# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sistema de Gestión de Subvenciones API
  description: |
    API REST para el Sistema de Gestión de Subvenciones de Alquiler.
    
    ## Autenticación
    La API utiliza JWT (JSON Web Tokens) para autenticación. Los tokens se envían en el header `Authorization` con el prefijo `Bearer`.
    
    ## Roles
    - **BENEFICIARIO**: Solicitante de la subvención
    - **ARRENDADOR**: Propietario del inmueble
    - **DIGER**: Personal administrativo que gestiona solicitudes
    - **DIRECTORA**: Directora que aprueba solicitudes
    - **ORDENADOR_GASTO**: Responsable de firma final
    - **CRI**: Centro de Recursos Informáticos (auditoría)
    
    ## Estados del Proceso
    - BORRADOR → ENVIADA → EN_VALIDACION → PENDIENTE_DIRECTORA → APROBADA_DIRECTORA → REVISION_ORDENADOR → APROBADA → CERRADA
    - También: RECHAZADA, CORRECCION_SOLICITADA
  version: 1.0.0
  contact:
    name: Soporte Técnico
    email: soporte@subvenciones.gob.es
  license:
    name: Propietario
servers:
  - url: http://localhost:4000/api
    description: Servidor de desarrollo
  - url: https://api.subvenciones.gob.es/api
    description: Servidor de producción
...
paths:
...
  /users:
    get:
      tags: [Usuarios]
      summary: Listar usuarios
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: search
          in: query
          schema:
            type: string
        - name: rol
          in: query
          schema:
            $ref: '#/components/schemas/Rol'
        - name: estado
          in: query
          schema:
            $ref: '#/components/schemas/EstadoUsuario'
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Usuarios]
      summary: Crear usuario
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
      responses:
        '201':
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/{id}:
    get:
      tags: [Usuarios]
      summary: Obtener usuario por ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '200':
          description: Usuario encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Usuarios]
      summary: Actualizar usuario
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/userId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUser'
      responses:
        '200':
          description: Usuario actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/{id}/deactivate:
    post:
      tags: [Usuarios]
      summary: Desactivar usuario
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '200':
          description: Usuario desactivado
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/{id}/unlock:
    post:
      tags: [Usuarios]
      summary: Desbloquear usuario
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '200':
          description: Usuario desbloqueado
        '403':
          $ref: '#/components/responses/Forbidden'

  # ==================== PROCESOS ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    userId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Solicitud inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: Sin permisos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
    Rol:
      type: string
      enum: [BENEFICIARIO, ARRENDADOR, DIGER, DIRECTORA, ORDENADOR_GASTO, CRI]
    EstadoUsuario:
      type: string
      enum: [ACTIVO, INACTIVO, BLOQUEADO]
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        cedula:
          type: string
        email:
          type: string
          format: email
        nombreCompleto:
          type: string
        rol:
          $ref: '#/components/schemas/Rol'
        estado:
          $ref: '#/components/schemas/EstadoUsuario'
        mfaHabilitado:
          type: boolean
        creadoEn:
          type: string
          format: date-time
    CreateUser:
      type: object
      required: [cedula, email, nombreCompleto, password, rol]
      properties:
        cedula:
          type: string
        email:
          type: string
          format: email
        nombreCompleto:
          type: string
        password:
          type: string
          minLength: 10
        rol:
          $ref: '#/components/schemas/Rol'
    UpdateUser:
      type: object
      properties:
        email:
          type: string
          format: email
        nombreCompleto:
          type: string
        rol:
          $ref: '#/components/schemas/Rol'
...
security:
  - bearerAuth: []